generate training answers with llama
