+++ lat/generation_utils.py
@@ -1,27 +1,7 @@
 import json
-import os
 import torch
-import jsonlines
-from tqdm import tqdm
-from dotenv import load_dotenv
 
-from lat.modeling import Llama7BChatHelper
 from lat.generate_steering_vectors import generate_single_steering_vector
-from lat.data import ComparisonDataset
-from lat.utils import system_prompt, data_path, jailbreaks_path
-
-load_dotenv()
-token = os.getenv("HF_TOKEN")
-system_prompt = "You are a helpful, honest and concise assistant."
-QUESTIONS = []
-
-file_path = 'datasets/refusal/filtered_questions.jsonl'
-
-# Open the JSONL file and extract questions.
-with jsonlines.open(file_path) as reader:
-    for item in reader:
-        if 'question' in item:
-            QUESTIONS.append(item)
 
 
 def get_vec(directory, layer, vec_type=""):
@@ -79,12 +59,21 @@ def generate_with_vector(model, questions, directory, saved_vector, example=None
                             layer, multiplier * vec.cuda())
 
                 # Batch questions
+                if "source" in questions[0]:
+                    source_exists = True
+                else:
+                    source_exists = False
                 for i in range(0, len(questions), batch_size):
                     # batch, and handle the case where we have less than batch_size questions
                     batched_questions = [
                         q["question"] for q in questions[i: min(i + batch_size, len(questions))]]
                     batched_categories = [
                         q["category"] for q in questions[i: min(i + batch_size, len(questions))]]
+                    if source_exists:
+                        batched_sources = [
+                            q["source"] for q in questions[i: min(i + batch_size, len(questions))]]
+                    else:
+                        batched_sources = None
                     generated_texts = model.generate_text(
                         batched_questions, max_new_tokens=max_new_tokens)
 
@@ -95,8 +84,14 @@ def generate_with_vector(model, questions, directory, saved_vector, example=None
                         print(f"Answer: {text}")
                         print(
                             f"Settings: layer {layer}, multiplier {multiplier}, directory {directory}, question_type {question_type}")
-                        answers.append(
-                            {"question": question, "answer": text, "category": category, "vec_name": vec_name})
+                        if batched_sources:
+                            source = batched_sources[batched_questions.index(
+                                question)]
+                            answers.append({"question": question, "answer": text,
+                                            "category": category, "source": source, "vec_name": vec_name})
+                        else:
+                            answers.append(
+                                {"question": question, "answer": text, "category": category, "vec_name": vec_name})
 
             layer_results.append(
                 {"multiplier": multiplier, "answers": answers})
@@ -104,37 +99,3 @@ def generate_with_vector(model, questions, directory, saved_vector, example=None
 
     with open(f"{directory}/{question_type}results.json", "w") as jfile:
         json.dump(all_results, jfile)
-
-
-if __name__ == "__main__":
-    model = Llama7BChatHelper(token, system_prompt, generation=True)
-    output_folder_name = "vanilla_steering"
-    os.makedirs(output_folder_name, exist_ok=True)
-    # generate_with_vector(model, QUESTIONS, output_folder_name, saved_vector=True)
-    # generate_with_vector(model, QUESTIONS, output_folder_name, saved_vector=False, question_type="vanilla_")
-    with open(data_path, "r") as f:
-        data = json.load(f)
-    dataset = ComparisonDataset(data, system_prompt)
-    # for i in range(10):
-    #     example = dataset[i]
-    #     output_folder_name = f"example{i}_steering"
-    #     os.makedirs(output_folder_name, exist_ok=True)
-    #     generate_with_vector(model, QUESTIONS, output_folder_name, saved_vector=False, example=example)
-
-    with open(jailbreaks_path, "r") as f:
-        jailbreaks = json.load(f)
-    for jailbreak in jailbreaks:
-        jailbreak_name = jailbreak["name"]
-        print(f"Generating for {jailbreak_name}")
-        jailbreak_prompt = jailbreak["prompt"]
-        jailbreak_questions = []
-        for question in QUESTIONS:
-            jailbreak_questions.append(
-                {"question": jailbreak_prompt + question["question"], "category": question["category"]})
-
-        output_folder_name = "vanilla_steering"
-        generate_with_vector(model, jailbreak_questions, output_folder_name,
-                             saved_vector=True, question_type=f"{jailbreak_name}_")
-        # comparison with doing nothing
-        generate_with_vector(model, jailbreak_questions, output_folder_name,
-                             saved_vector=False, question_type=f"vanilla_{jailbreak_name}_")
